refactor: add validateions in budie helper
